+++ apps/web/app/future/team/[slug]/[type]/embed/page.tsx
@@ -1,10 +1,12 @@
-import LegacyPage, { getServerSideProps } from "@pages/team/[slug]/[type]/embed";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import withEmbedSsrAppDir from "app/WithEmbedSSR";
 import { WithLayout } from "app/layoutHOC";
 
-const getData = withAppDirSsr(getServerSideProps);
+import { getServerSideProps, type PageProps } from "@lib/team/[slug]/[type]/getServerSideProps";
+
+import LegacyPage from "@components/pages/team/[slug]/[type]";
+
+const getData = withAppDirSsr<PageProps>(getServerSideProps);
 const getEmbedData = withEmbedSsrAppDir(getData);
 
-// @ts-expect-error Type is missing the following properties from type: entity, duration, booking, away, and 7 more.
 export default WithLayout({ getLayout: null, getData: getEmbedData, Page: LegacyPage })<"P">;
