+++ apps/web/app/future/team/[slug]/embed/page.tsx
@@ -1,7 +1,10 @@
-import LegacyPage, { type PageProps, getServerSideProps } from "@pages/team/[slug]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import withEmbedSsrAppDir from "app/WithEmbedSSR";
 import { WithLayout } from "app/layoutHOC";
+
+import { getServerSideProps, type PageProps } from "@lib/team/[slug]/getServerSideProps";
+
+import LegacyPage from "@components/pages/team/[slug]";
 
 const getData = withAppDirSsr<PageProps>(getServerSideProps);
 const getEmbedData = withEmbedSsrAppDir(getData);
