+++ apps/web/pages/team/[slug]/embed.tsx
@@ -1,9 +1,7 @@
-"use client";
-
 import withEmbedSsr from "@lib/withEmbedSsr";
 
 import { getServerSideProps as _getServerSideProps } from "../../team/[slug]";
 
-export { default } from "../../team/[slug]";
+export { default, type PageProps } from "../../team/[slug]";
 
 export const getServerSideProps = withEmbedSsr(_getServerSideProps);
