+++ apps/web/pages/teams/index.tsx
@@ -1,43 +1,16 @@
-"use client";
-
 import { getLayout } from "@calcom/features/MainLayout";
-import { TeamsListing } from "@calcom/features/ee/teams/components";
-import { ShellMain } from "@calcom/features/shell/Shell";
-import { WEBAPP_URL } from "@calcom/lib/constants";
-import { useLocale } from "@calcom/lib/hooks/useLocale";
-import { trpc } from "@calcom/trpc/react";
-import { Button } from "@calcom/ui";
-import { Plus } from "@calcom/ui/components/icon";
 
-import PageWrapper from "@components/PageWrapper";
+import { type AppProps } from "@lib/app-providers";
+
+import PageWrapper, { type CalPageWrapper } from "@components/PageWrapper";
+import TeamsPage from "@components/pages/teams";
 
 export { getServerSideProps } from "@lib/teams/getServerSideProps";
 
-function Teams() {
-  const { t } = useLocale();
-  const [user] = trpc.viewer.me.useSuspenseQuery();
-
-  return (
-    <ShellMain
-      heading={t("teams")}
-      hideHeadingOnMobile
-      subtitle={t("create_manage_teams_collaborative")}
-      CTA={
-        (!user.organizationId || user.organization.isOrgAdmin) && (
-          <Button
-            data-testid="new-team-btn"
-            variant="fab"
-            StartIcon={Plus}
-            type="button"
-            href={`${WEBAPP_URL}/settings/teams/new?returnTo=${WEBAPP_URL}/teams`}>
-            {t("new")}
-          </Button>
-        )
-      }>
-      <TeamsListing />
-    </ShellMain>
-  );
-}
+const Teams = TeamsPage as unknown as CalPageWrapper & {
+  getLayout: AppProps["Component"]["getLayout"];
+  requiresLicense: AppProps["Component"]["requiresLicense"];
+};
 
 Teams.requiresLicense = false;
 Teams.PageWrapper = PageWrapper;
